refactor: move google maps into a separate reusable component
