+++ frontend/src/App.scss
@@ -31,3 +31,16 @@ main {
     padding: 1em;
   }
 }
+
+.widget {
+  padding: 2em;
+  border: 1px solid #c8c8c8;
+  border-radius: 1em;
+  min-width: 300px;
+  overflow: hidden;
+
+  h3 {
+    font-size: 1.5em;
+    margin-bottom: 1em;
+  }
+}
