--- frontend/src/components/Home/MapWidget/MapWidget.scss
@@ -1,16 +1,0 @@
-.map-widget {
-  padding: 0;
-
-  .airport-map {
-    width: 100%;
-    min-height: 400px;
-    height: 100%;
-
-    .map-label {
-      padding: 0.4em 0.6em;
-      border: 1px solid lightgray;
-      border-radius: 0.5em;
-      background-color: #222222;
-    }
-  }
-}
