+++ frontend/src/components/Home/NearbyWidget/NearbyWidget.tsx
@@ -22,7 +22,7 @@ const AirportItem = ({ name = '', iata = '', location = '', distance = -1 }) => 
 
 const NearbyWidget = ({ airports = [] }: { airports: AirportEntity[] }) => {
   return (
-    <div className='home-widget'>
+    <div className='widget'>
       <h3>Airports near you location</h3>
       <ul className='airport-list'>
       { airports.map(({ name, iata, location, distance }) => <AirportItem key={iata} { ...{ name, iata, location, distance }} /> ) }
