+++ frontend/src/components/Map/Map.scss
@@ -1,0 +1,24 @@
+.map-widget {
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  .airport-map {
+    width: 100%;
+    min-height: 400px;
+    height: 100%;
+
+    .map-label {
+      padding: 0.4em 0.6em;
+      border: 1px solid lightgray;
+      border-radius: 0.5em;
+      background-color: #222222;
+    }
+  }
+
+  .map-status {
+    text-align: center;
+  }
+}
